🎨 Rename pages path
