+++ src/pages/Animal/Registration/index.tsx
