+++ src/pages/Animal/Registration/styles.ts
