+++ src/pages/Animal/Success/index.tsx
