+++ src/pages/Animal/Success/styles.ts
