+++ src/routes.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
 
 // Importing our pages.
-import AnimalRegistration from './pages/AnimalRegistration/Signin';
-import AnimalRegistrationSuccess from './pages/AnimalRegistration/Success';
+import AnimalRegistration from './pages/Animal/Registration';
+import AnimalRegistrationSuccess from './pages/Animal/Success';
 import Home from './pages/Home';
 import Login from './pages/Login';
 import Registration from './pages/Registration';
